requirements.txt
boto3>=1.42.24
dbt-core>=1.11.2
dbt-postgres>=1.10.0
fastapi>=0.128.0
pandas>=2.3.3
psycopg2-binary>=2.9.11
pyarrow>=22.0.0
pydantic>=2.12.5
pyjwt>=2.10.1
python-dotenv>=1.2.1
requests>=2.32.5
sqlalchemy>=2.0.45
uvicorn>=0.40.0